API/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.38.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.1.0
fastapi==0.115.8
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
ipython==8.29.0
jedi==0.19.1
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.17
langchain-anthropic==0.2.4
langchain-core==0.3.33
langchain-google-genai==2.0.4
langchain-text-splitters==0.3.5
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.1.139
matplotlib-inline==0.1.7
msgpack==1.1.0
multidict==6.1.0
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.11
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.45.3
tenacity==9.0.0
tokenizers==0.20.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
yarl==1.17.1
